Pablo

+++ src/components/ProfileComponents/EducationProvider/EducationProvider.jsx
@@ -1,6 +1,7 @@
 import { useSelector, useDispatch } from "react-redux";
 import { useState, useEffect } from "react";
 import { deleteService } from "../../../redux/actions/index";
+import Swal from "sweetalert2";
 import Form from "../../Form/FormEducation/FormEducation";
 import style from "./EducationProvider.module.sass";
 
@@ -48,7 +49,22 @@ function EducationProvider() {
       idPeople: infoUserLog.idPeople,
       idOption: idOption,
     };
-    dispatch(deleteService(deleteData));
+
+    Swal.fire({
+      title: "Quieres eliminar este nivel de Educación?",
+      text: `Click en Aceptar para eliminarlo, o dale a Cancelar para regresar`,
+      footer: "Confirma que quieres eliminar la Educación seleccionado",
+      icon: "alert",
+      showDenyButton: true,
+      denyButtonText: "Cancelar",
+      denyButtonColor: "Grey",
+      confirmButtonText: "Eliminar",
+      confirmButtonColor: "Red",
+    }).then((response) => {
+      if (response.isConfirmed) {
+        dispatch(deleteService(deleteData));
+      } 
+    });
   };
 
   return (
